Universe framework: make bootstrap step configurable
