+++ universe/universe-core/src/main/java/org/corfudb/universe/universe/group/cluster/corfu/AbstractCorfuCluster.java
@@ -2,8 +2,10 @@ package org.corfudb.universe.universe.group.cluster.corfu;
 
 import com.google.common.collect.ImmutableSortedSet;
 import lombok.extern.slf4j.Slf4j;
+import org.corfudb.runtime.BootstrapUtil;
 import org.corfudb.runtime.CorfuRuntime;
 import org.corfudb.runtime.CorfuRuntime.CorfuRuntimeParameters.CorfuRuntimeParametersBuilder;
+import org.corfudb.runtime.view.Layout;
 import org.corfudb.universe.api.common.LoggingParams;
 import org.corfudb.universe.api.deployment.DeploymentParams;
 import org.corfudb.universe.api.universe.UniverseParams;
@@ -11,6 +13,8 @@ import org.corfudb.universe.api.universe.group.cluster.AbstractCluster;
 import org.corfudb.universe.api.universe.node.ApplicationServer;
 import org.corfudb.universe.universe.node.client.LocalCorfuClient;
 import org.corfudb.universe.universe.node.server.corfu.CorfuServerParams;
+
+import java.time.Duration;
 
 @Slf4j
 public abstract class AbstractCorfuCluster<
@@ -49,4 +53,25 @@ public abstract class AbstractCorfuCluster<
 
     protected abstract ImmutableSortedSet<String> getClusterLayoutServers();
 
+    @Override
+    public void bootstrap() {
+        bootstrap(params.getBootstrapParams().isBootstrapEnabled());
+    }
+
+    @Override
+    public void bootstrap(boolean enabled) {
+        if (!enabled) {
+            log.warn("Bootstrap disabled for: {}", params.getName());
+            return;
+        }
+
+        Layout layout = buildLayout();
+        log.info("Bootstrap docker corfu cluster. Cluster: {}. layout: {}", params.getName(), layout.asJSONString());
+
+        int retries = params.getBootstrapParams().getBootStrapRetries();
+        Duration duration = params.getBootstrapParams().getRetryDuration();
+        BootstrapUtil.bootstrap(layout, retries, duration);
+    }
+
+    protected abstract Layout buildLayout();
 }
