+++ universe/universe-infrastructure/src/main/java/org/corfudb/universe/infrastructure/docker/universe/group/cluster/DockerMangleCluster.java
@@ -50,6 +50,11 @@ public class DockerMangleCluster extends AbstractCluster<
     }
 
     @Override
+    public void bootstrap(boolean force) {
+        // NOOP
+    }
+
+    @Override
     protected DockerMangleServer buildServer(
             DockerContainerParams<MangleServerParams> deploymentParams) {
