+++ universe/universe-infrastructure/src/main/java/org/corfudb/universe/infrastructure/docker/universe/group/cluster/DockerPrometheusCluster.java
@@ -49,6 +49,11 @@ public class DockerPrometheusCluster extends AbstractCluster<
     }
 
     @Override
+    public void bootstrap(boolean force) {
+        // NOOP
+    }
+
+    @Override
     protected DockerPrometheusServer buildServer(DockerContainerParams<PromServerParams> deploymentParams) {
 
         DockerManager<PromServerParams> dockerManager = DockerManager
